requirements.txt
bs4
python-dotenv
gradio
markdown
openai
langchain
